04_assignments/Assignments 01/03_online_project/qr-code-scanner/pyproject.toml
[project]
name = "qr-code-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "qrcode>=8.1",
]
